split out layout module
