+++ src/pages/auth/error.tsx
@@ -1,4 +1,4 @@
-import { Error } from "@/components/Error";
+import { Error } from "@/components/error";
 import { getQueryParam } from "@/utils/query";
 import type { NextPageWithLayout } from "@/utils/type";
 import { useRouter } from "next/router";
